+++ serialization/src/commonTest/kotlin/nl/adaptivity/xml/serialization/regressions/TestCachedDescriptorCache.kt
@@ -21,10 +21,24 @@
 package nl.adaptivity.xml.serialization.regressions
 
 import io.github.pdvrieze.xmlutil.testutil.assertXmlEquals
+import kotlinx.serialization.Polymorphic
 import kotlinx.serialization.Serializable
+import kotlinx.serialization.decodeFromString
 import kotlinx.serialization.encodeToString
+import kotlinx.serialization.modules.SerializersModule
+import kotlinx.serialization.modules.polymorphic
+import kotlinx.serialization.modules.subclass
+import nl.adaptivity.xmlutil.serialization.DefaultFormatCache
+import nl.adaptivity.xmlutil.serialization.OutputKind
 import nl.adaptivity.xmlutil.serialization.XML
+import nl.adaptivity.xmlutil.serialization.XmlValue
+import nl.adaptivity.xmlutil.serialization.structure.XmlCompositeDescriptor
+import nl.adaptivity.xmlutil.serialization.structure.XmlListDescriptor
+import nl.adaptivity.xmlutil.serialization.structure.XmlPolymorphicDescriptor
+import nl.adaptivity.xmlutil.serialization.structure.XmlPrimitiveDescriptor
 import kotlin.test.Test
+import kotlin.test.assertEquals
+import kotlin.test.assertIs
 
 class TestCachedDescriptorCache {
 
@@ -39,12 +53,64 @@ class TestCachedDescriptorCache {
         assertXmlEquals("<Outer><Inner2 data3=\"a\" data4=\"b\" data5=\"c\"/></Outer>", serialized2)
     }
 
+    @Test
+    fun testCacheXmlValues() {
+        val format = XML(listModule) {
+            autoPolymorphic = true
+            recommended_0_91_0 {
+                formatCache = DefaultFormatCache() // skip the layering for debugging
+            }
+        }
+
+        val desc1 = assertIs<XmlCompositeDescriptor>(format.xmlDescriptor(OuterList1.serializer()).getElementDescriptor(0))
+        val desc2 = assertIs<XmlCompositeDescriptor>(format.xmlDescriptor(OuterList2.serializer()).getElementDescriptor(0))
+
+
+        val deserialized1 = format.decodeFromString<OuterList1>("<OuterList1><string xmlns=\"http://www.w3.org/2001/XMLSchema\">text</string></OuterList1>")
+
+        assertEquals(1, desc1.elementsCount)
+        val desc1list = assertIs<XmlListDescriptor>(desc1.getElementDescriptor(0))
+        val desc1data = assertIs<XmlPolymorphicDescriptor>(desc1list.getElementDescriptor(0))
+        assertEquals(2, desc1data.polyInfo.size)
+        val desc1string = assertIs<XmlPrimitiveDescriptor>(desc1data.polyInfo["kotlin.String"])
+        val desc1inner = assertIs<XmlCompositeDescriptor>(desc1data.polyInfo["nl.adaptivity.xml.serialization.regressions.TestCachedDescriptorCache.Inner1"])
+        assertEquals(OutputKind.Element, desc1string.outputKind)
+
+        assertEquals(listOf("text"), deserialized1.data)
+
+        assertEquals(1, desc2.elementsCount)
+        val desc2list = assertIs<XmlListDescriptor>(desc2.getElementDescriptor(0))
+        val desc2data = assertIs<XmlPolymorphicDescriptor>(desc2list.getElementDescriptor(0))
+        assertEquals(2, desc2data.polyInfo.size)
+        val desc2string = assertIs<XmlPrimitiveDescriptor>(desc2data.polyInfo["kotlin.String"])
+        val desc2inner = assertIs<XmlCompositeDescriptor>(desc2data.polyInfo["nl.adaptivity.xml.serialization.regressions.TestCachedDescriptorCache.Inner1"])
+        assertEquals(OutputKind.Mixed, desc2string.outputKind)
+
+        val deserialized2 = format.decodeFromString<OuterList2>("<OuterList>text</OuterList>")
+        assertEquals(listOf("text"), deserialized2.data)
+    }
+
     @Serializable
     data class Outer<T>(val data: T)
+
+    @Serializable
+    data class OuterList1(val data: List<@Polymorphic Any>)
+
+    @Serializable
+    data class OuterList2(@XmlValue val data: List<@Polymorphic Any>)
 
     @Serializable
     data class Inner1(val data1: Int, val data2: Int)
 
     @Serializable
     data class Inner2(val data3: String, val data4: String, val data5: String)
+
+    companion object {
+        val listModule = SerializersModule {
+            polymorphic(Any::class) {
+                subclass(Inner1::class)
+                subclass(String::class)
+            }
+        }
+    }
 }